+++ lib/sidekiq/processor.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 require 'sidekiq/util'
 require 'sidekiq/fetch'
+require 'sidekiq/job_logger'
+require 'sidekiq/job_retry'
 require 'thread'
 require 'concurrent/map'
 require 'concurrent/atomic/atomic_fixnum'
@@ -37,7 +39,8 @@ module Sidekiq
       @thread = nil
       @strategy = (mgr.options[:fetch] || Sidekiq::BasicFetch).new(mgr.options)
       @reloader = Sidekiq.options[:reloader]
-      @executor = Sidekiq.options[:executor]
+      @logging = Sidekiq::JobLogger.new
+      @retrier = Sidekiq::JobRetry.new
     end
 
     def terminate(wait=false)
@@ -116,30 +119,59 @@ module Sidekiq
       nil
     end
 
+    def dispatch(job_hash, queue)
+      # since middleware can mutate the job hash
+      # we clone here so we report the original
+      # job structure to the Web UI
+      pristine = cloned(job_hash)
+
+      @retrier.global(job_hash, queue) do
+        @logging.call(job_hash, queue) do
+          stats(pristine, queue) do
+            # Rails 5 requires a Reloader to wrap code execution.  In order to
+            # constantize the worker and instantiate an instance, we have to call
+            # the Reloader.  It handles code loading, db connection management, etc.
+            # Effectively this block denotes a "unit of work" to Rails.
+            @reloader.call do
+              klass  = job_hash['class'.freeze].constantize
+              worker = klass.new
+              worker.jid = job_hash['jid'.freeze]
+              @retrier.local(worker, job_hash, queue) do
+                yield worker
+              end
+            end
+          end
+        end
+      end
+    end
+
     def process(work)
       jobstr = work.job
       queue = work.queue_name
 
       ack = false
       begin
-        job_hash = Sidekiq.load_json(jobstr)
-        @reloader.call do
-          klass  = job_hash['class'.freeze].constantize
-          worker = klass.new
-          worker.jid = job_hash['jid'.freeze]
-
-          stats(worker, job_hash, queue) do
-            Sidekiq.server_middleware.invoke(worker, job_hash, queue) do
-              @executor.call do
-                # Only ack if we either attempted to start this job or
-                # successfully completed it. This prevents us from
-                # losing jobs if a middleware raises an exception before yielding
-                ack = true
-                execute_job(worker, cloned(job_hash['args'.freeze]))
-              end
-            end
-          end
+        # Treat malformed JSON like a process crash -- don't acknowledge it.
+        # * In Sidekiq, the error will be logged but job discarded.
+        # * In Sidekiq Pro, the error will be logged and the job retried when
+        #   it is recovered by the reliability algorithm.  The job may act like
+        #   a poison pill and never execute until manually removed but job loss
+        #   is considered worse.
+        job_hash = nil
+        begin
+          job_hash = Sidekiq.load_json(jobstr)
+        rescue => ex
+          Sidekiq.logger.error { "Pushing job to dead queue due to invalid JSON: #{ex}" }
+          send_to_morgue(jobstr)
           ack = true
+          raise
+        end
+
+        ack = true
+        dispatch(job_hash, queue) do |worker|
+          Sidekiq.server_middleware.invoke(worker, job_hash, queue) do
+            execute_job(worker, cloned(job_hash['args'.freeze]))
+          end
         end
       rescue Sidekiq::Shutdown
         # Had to force kill this job because it didn't finish
@@ -151,6 +183,17 @@ module Sidekiq
         raise
       ensure
         work.acknowledge if ack
+      end
+    end
+
+    def send_to_morgue(msg)
+      now = Time.now.to_f
+      Sidekiq.redis do |conn|
+        conn.multi do
+          conn.zadd('dead', now, msg)
+          conn.zremrangebyscore('dead', '-inf', now - DeadSet.timeout)
+          conn.zremrangebyrank('dead', 0, -DeadSet.max_jobs)
+        end
       end
     end
 
@@ -166,9 +209,9 @@ module Sidekiq
     PROCESSED = Concurrent::AtomicFixnum.new
     FAILURE = Concurrent::AtomicFixnum.new
 
-    def stats(worker, job_hash, queue)
+    def stats(job_hash, queue)
       tid = thread_identity
-      WORKER_STATE[tid] = {:queue => queue, :payload => cloned(job_hash), :run_at => Time.now.to_i }
+      WORKER_STATE[tid] = {:queue => queue, :payload => job_hash, :run_at => Time.now.to_i }
 
       begin
         yield
@@ -184,8 +227,8 @@ module Sidekiq
     # Deep clone the arguments passed to the worker so that if
     # the job fails, what is pushed back onto Redis hasn't
     # been mutated by the worker.
-    def cloned(ary)
-      Marshal.load(Marshal.dump(ary))
+    def cloned(thing)
+      Marshal.load(Marshal.dump(thing))
     end
 
   end