+++ lib/sidekiq/rails.rb
@@ -27,37 +27,9 @@ module Sidekiq
       #
       Sidekiq.configure_server do |_|
         if ::Rails::VERSION::MAJOR >= 5
-          # The reloader also takes care of ActiveRecord but is incompatible with
-          # the ActiveRecord middleware so make sure it's not in the chain already.
-          if defined?(Sidekiq::Middleware::Server::ActiveRecord) && Sidekiq.server_middleware.exists?(Sidekiq::Middleware::Server::ActiveRecord)
-            raise ArgumentError, "You are using the Sidekiq ActiveRecord middleware and the new Rails 5 reloader which are incompatible. Please remove the ActiveRecord middleware from your Sidekiq middleware configuration."
-          elsif ::Rails.application.config.cache_classes
-            # The reloader API has proven to be troublesome under load in production.
-            # We won't use it at all when classes are cached, see #3154
-            Sidekiq.logger.debug { "Autoload disabled in #{::Rails.env}, Sidekiq will not reload changed classes" }
-            Sidekiq.options[:executor] = Sidekiq::Rails::Executor.new
-          else
-            Sidekiq.logger.debug { "Enabling Rails 5+ live code reloading, so hot!" }
-            Sidekiq.options[:reloader] = Sidekiq::Rails::Reloader.new
-            Psych::Visitors::ToRuby.prepend(Sidekiq::Rails::PsychAutoload)
-          end
+          Sidekiq.options[:reloader] = Sidekiq::Rails::Reloader.new
+          Psych::Visitors::ToRuby.prepend(Sidekiq::Rails::PsychAutoload)
         end
-      end
-    end
-
-    class Executor
-      def initialize(app = ::Rails.application)
-        @app = app
-      end
-
-      def call
-        @app.executor.wrap do
-          yield
-        end
-      end
-
-      def inspect
-        "#<Sidekiq::Rails::Executor @app=#{@app.class.name}>"
       end
     end